Fix up espressoBuildConfig capability
The espressoBuildConfig capability is supposed to be able to just be a json string (see release notes: https://github.com/appium/appium/releases/tag/v1.19.0 ) however it doesn't appear to work properly in the dotnet driver.  This adds a check for the capability and tests if it's a file or json, and if json, writes it out to a temporary file and specifies the file path instead (which does work as expected).